requirements.txt
fastapi
uvicorn
jinja2
python-multipart
snowflake-connector-python
python-dotenv